Fixing array access. Added Ruby compiler using Ramaze. AIR app to run scripts. Had a bug in array access that I didn't notice since I didn't copy the test code into the app properly. Added a web-based Ruby compiler using Ramaze. Failed to get Rails or Camping running on Ruby 1.9. Didn't feel like shelling out from Ruby 1.8 to 1.9 to do the compile, although that might be the only option if I want to run this on my web server. Updated AIR app to have a TextArea to enter Ruby code into and execute, by talking to the Ramaze app. It has some hardcoded local paths to my HD now.
Fixed oversight in variable assignment mapping from Ruby to AS3. Needed
to detect var= style method calls and do assignment instead.

Fixed oversight in AS3 -> Ruby conversion method that was using == to
test if a value was == true, which passes for many values. Switched to
=== and maybe should reorder the tests as well to be safe.

duparray is an odd bytecode that asks the compiler to essentially drop
the array from the opcode onto the stack. Not sure how this would work
if the array contains local variable access. Assuming that the bytecodes
would compile differently for [1,true,nil] and [a,b,c].